Day 2 part 2 update
on day 2 part 2 I had to use chat gpt because I could't figure out how to do the Problem Dampener which allowed us to skip a level in the report and go on to the next. i learned to use a colon (:) in front of my i variable so that it took all levels in my report before the variable as well as after the variable. it tests all the variables and returns true if there are no issues and returns false if there is an unsafe issue. because we replace modified report with report, but we use the definition of is_safe, increasing and decreasing require all() iterations to function